requirements.txt
python-telegram-bot>=20.0
google-generativeai>=0.7.0
aiohttp>=3.9.0